move youtube panel to settings
